Move label color map
